refactor(surrogate): move sabi.posterior -> sabi.surrogate (#7)
Move the package directory and update all imports + the docs/design.md
repo-layout sketch. The package is the home of the surrogate
machinery (SurrogateDistribution, WeightedEmpiricalRandomMeasure,
expected_target estimator, pushforward primitives) — `sabi.surrogate`
reads more directly than `sabi.posterior` after the SurrogatePosterior
rename. No behavior changes.
