--- src/sabi/posterior/__init__.py
@@ -1,9 +1,0 @@
-from sabi.posterior.estimators import expected_target
-from sabi.posterior.surrogate_distribution import SurrogateDistribution
-from sabi.posterior.weighted_empirical import WeightedEmpiricalRandomMeasure
-
-__all__ = [
-    "SurrogateDistribution",
-    "WeightedEmpiricalRandomMeasure",
-    "expected_target",
-]
