+++ src/sabi/surrogate/__init__.py
@@ -1,0 +1,9 @@
+from sabi.surrogate.estimators import expected_target
+from sabi.surrogate.surrogate_distribution import SurrogateDistribution
+from sabi.surrogate.weighted_empirical import WeightedEmpiricalRandomMeasure
+
+__all__ = [
+    "SurrogateDistribution",
+    "WeightedEmpiricalRandomMeasure",
+    "expected_target",
+]
