+++ src/sabi/surrogate/_dirac.py
