+++ src/sabi/surrogate/estimators.py
@@ -37,8 +37,8 @@ from probpipe.core._distribution_base import Distribution
 from probpipe.core._numeric_record_distribution import NumericRecordDistribution
 from probpipe.core.constraints import Constraint
 
-from sabi.posterior.surrogate_distribution import SurrogateDistribution
-from sabi.posterior.weighted_empirical import WeightedEmpiricalRandomMeasure
+from sabi.surrogate.surrogate_distribution import SurrogateDistribution
+from sabi.surrogate.weighted_empirical import WeightedEmpiricalRandomMeasure
 from sabi.problems.forms import LogDensityForm
 from sabi.emulators.base import Emulator
